Fix OOB read on metadata IN path (metadataInIdx size_t->int) + test
VEC0_METADATA_OPERATOR_IN declared `size_t metadataInIdx = -1`, which wraps
to SIZE_MAX. The subsequent `if (metadataInIdx < 0)` not-found guard is dead
code for an unsigned type, so a missing entry fell through to an out-of-bounds
read indexing the metadata-IN array. The parallel path already uses `int`.

+++ sqlite-vec.c
@@ -6987,7 +6987,7 @@ int vec0_metadata_filter_text(vec0_vtab * p, sqlite3_value * value, const void *
     }
 
     case VEC0_METADATA_OPERATOR_IN: {
-      size_t metadataInIdx = -1;
+      int metadataInIdx = -1;
       for(size_t i = 0; i < aMetadataIn->length; i++) {
         struct Vec0MetadataIn * metadataIn = &(((struct Vec0MetadataIn *) aMetadataIn->z)[i]);
         if(metadataIn->argv_idx == argv_idx) {
